OPENEUROPA-2340: Conflict resolving.
